docker: Install specific version of Jinja2 templating library
When regenerating docker image from scratch (i.e. not using previously
built cached images), the build would eventually fail with an error:

```
Traceback (most recent call last):
  File "build/templates.py", line 68, in <module>
    import jinja2
  File "/usr/local/lib/python3.5/dist-packages/jinja2/__init__.py", line 5, in <module>
    from .bccache import BytecodeCache
  File "/usr/local/lib/python3.5/dist-packages/jinja2/bccache.py", line 61
    self.code: t.Optional[CodeType] = None
             ^
SyntaxError: invalid syntax
```

The issue comes from the fact that installing Jinja2 Python library
downloads the latest version, which has been updated to version 3.0.0.
The changelog for this version says:
- Drop support for Python 2.7 and 3.5.

There is the same issue with MarkupSafe - a dependency of Jinja2.

Since we are using Python 3.5 inside the docker image, we now pin the
versions of Jinja2 and MarkupSafe to the latest versions below 3.0.0
and 2.0.0 respectively.